+++ CHANGELOG.md
@@ -1,16 +1,24 @@
 # Changelog
 
+## v0.3.0 2020-01-21
+
+- add types/index.d.ts
+- update eslint/prettier/...
+
 ## v0.2.0 2018-08-26
 
 - use prettier
 - use async/await, require node >= 7.6.0
 
 ## v0.1.0 2016-06-02
+
 - use `promise.timeout@0.1.0`
 
 ## v0.0.2 2016-06-01
+
 - use `promise.timeout@0.0.3`, add `onCancel` support
 - use esformatter, rm jsbeautify
 
 ## v0.0.1 2016-05-16
+
 - first release